Create ScanStep component
